ai/requirements-m2.txt
# ── Core ML / PyTorch (Apple Silicon) ────────────────────────────────────────
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# ── MPS-compatible inference ──────────────────────────────────────────────────
onnxruntime-silicon          # CoreML Execution Provider for ONNX models
coremltools>=8.0

# ── API framework ─────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart             # multipart/form-data uploads
httpx>=0.27.0

# ── Data validation ───────────────────────────────────────────────────────────
pydantic>=2.9.0
pydantic-settings>=2.6.0

# ── Job queue ─────────────────────────────────────────────────────────────────
redis>=5.2.0

# ── Storage ───────────────────────────────────────────────────────────────────
boto3>=1.35.0

# ── 3D mesh ───────────────────────────────────────────────────────────────────
trimesh>=4.5.0
pyrender>=0.1.45
smplx>=0.1.28               # SMPL-X parametric body model API

# ── Body mesh (primary) — install separately ─────────────────────────────────
# HMR2.0 / 4D-Humans:
#   pip install git+https://github.com/shubham-goel/4D-Humans.git
# Download checkpoint: python -m hmr2.utils.download_models
# SMPL-X model weights (requires registration):
#   https://smpl-x.is.tue.mpg.de/ → SMPLX_NEUTRAL.npz → ~/smplx_models/

# ── Face identity (primary) — install separately ──────────────────────────────
# DECA:
#   pip install git+https://github.com/YadiraF/DECA.git
# DECA downloads its own FLAME weights on first run.

# ── Computer vision ───────────────────────────────────────────────────────────
opencv-python-headless==4.8.1.78   # pinned: 4.9+ requires numpy>=2, incompatible with smplx
numpy<2                             # smplx/mediapipe require numpy 1.x
mediapipe==0.10.9       # 0.10.10+ drops mp.solutions — keep pinned
Pillow>=11.0.0
scikit-image>=0.24.0        # SSIM for texture quality validation

# ── Video decode ──────────────────────────────────────────────────────────────
ffmpeg-python>=0.2.0
# brew install ffmpeg  (required separately)

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy>=1.26.0
scipy>=1.14.0
psutil>=6.1.0
python-dotenv>=1.0.0
coloredlogs>=15.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-timeout>=2.3.0

# ── Required env vars ─────────────────────────────────────────────────────────
# PYTORCH_ENABLE_MPS_FALLBACK=1   (set in .env or shell profile)
# BODY_MESH_MODEL=hmr2            (default)
# FACE_ID_MODEL=deca              (default)
# SEGMENTATION_MODEL=rembg        (default)
# SMPLX_MODEL_DIR=~/smplx_models  (default)
